cleanup from feedback
