+++ pkg/appimport/appimport_test.go
@@ -1,7 +1,6 @@
 package appimport
 
 import (
-	"fmt"
 	"path"
 	"testing"
 
@@ -17,40 +16,25 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
-const netName = "ddev_default"
-
-var (
-	testArchivePath = path.Join(testcommon.CreateTmpDir("appimport"), "db.tar.gz")
-	cwd             string
-)
-
-func TestMain(m *testing.M) {
-	testFile, err := os.Create(testArchivePath)
-	if err != nil {
-		log.Fatalf("failed to create test file: %v", err)
-	}
-	err = testFile.Close()
-	if err != nil {
-		log.Fatalf("failed to create test file: %v", err)
-	}
-
-	cwd, err = os.Getwd()
-	if err != nil {
-		log.Fatalf("failed to get cwd: %s", err)
-	}
-
-	fmt.Println("Running tests.")
-	testRun := m.Run()
-
-	err = os.RemoveAll(path.Dir(testArchivePath))
-	util.CheckErr(err)
-
-	os.Exit(testRun)
-}
-
 // TestValidateAsset tests validation of asset paths.
 func TestValidateAsset(t *testing.T) {
 	assert := assert.New(t)
+
+	testArchivePath := path.Join(testcommon.CreateTmpDir("appimport"), "db.tar.gz")
+
+	testFile, err := os.Create(testArchivePath)
+	if err != nil {
+		log.Fatalf("failed to create dummy test file: %v", err)
+	}
+	err = testFile.Close()
+	if err != nil {
+		log.Fatalf("failed to create dummy test file: %v", err)
+	}
+
+	cwd, err := os.Getwd()
+	if err != nil {
+		log.Fatalf("failed to get cwd: %s", err)
+	}
 
 	// test tilde expansion
 	userDir, err := homedir.Dir()
@@ -86,4 +70,7 @@ func TestValidateAsset(t *testing.T) {
 	_, err = ValidateAsset("appimport.go", "files")
 	assert.Error(err)
 	assert.Contains(err.Error(), "provided path is not a directory or archive")
+
+	err = os.RemoveAll(path.Dir(testArchivePath))
+	util.CheckErr(err)
 }
